chore(opentelemetry source): Refactor source into multiple files (#13658)
* chore(opentelemetry source): Move tests into separate module


* chore(opentelemetry source): split config from log service


* +pr feedback, clippy fix
